Feature/entries table animation

+++ src/lib/api.ts
@@ -44,7 +44,7 @@ export const getCollectionBySlug = async (slug: string) => {
     categories (slug),
     entries:entries!entries_collection_id_fkey (id, clue, answer),
     entries_count:entries!entries_collection_id_fkey(count)
-  `
+  `,
       )
       .eq("slug", slug)
       .maybeSingle();
@@ -67,7 +67,7 @@ export const getCollectionsByCategorySlug = async (slug: string) => {
     name,
     description,
     collections (id, slug, name, description)
-  `
+  `,
       )
       .eq("slug", slug)
       .maybeSingle();
@@ -79,12 +79,16 @@ export const getCollectionsByCategorySlug = async (slug: string) => {
   }
 };
 
-export const getLearnedEntries = async (userId: string) => {
+export const getLearnedEntries = async (
+  userId: string,
+  collectionId: number,
+) => {
   try {
     const { data, error } = await supabase
       .from("learned_entries")
       .select("entry_id,collection_id")
-      .eq("user_id", userId);
+      .eq("user_id", userId)
+      .eq("collection_id", collectionId);
     if (error) console.error(error.message);
     return data;
   } catch (error) {
@@ -94,9 +98,9 @@ export const getLearnedEntries = async (userId: string) => {
 };
 
 export const insertLearnedEntry = async (
-  entryId: number,
   collection_id: number,
-  userId: string
+  entryId: number,
+  userId: string,
 ) => {
   try {
     const { error } = await supabase.from("learned_entries").insert({
@@ -112,15 +116,17 @@ export const insertLearnedEntry = async (
 };
 
 export const deleteLearnedEntry = async (
+  collectionId: number,
   entryId: number,
-  collectionId: number
+  userId: string,
 ) => {
   try {
     const { error } = await supabase
       .from("learned_entries")
       .delete()
       .eq("entry_id", entryId)
-      .eq("collection_id", collectionId);
+      .eq("collection_id", collectionId)
+      .eq("user_id", userId);
     if (error) console.error(error.message);
   } catch (error) {
     console.error(`An unexpected error occurred: ${error}`);
